Some fixes to the SliderPanel
Fixed MinValue. I was able to go under it while dragging.
Fixed SliderX. It was innacurate from my testings and it was ignoring MinValue. (with some numbers wasn't reaching the end or it was going past it)
With these two above, you are also able to use negative numbers while its being dragged/hovered.
Added ItemOffset because panels were in the same place without it.
Commented prints. They are not needed if you are not debugging.
Added sound while its being dragged/hovered (took from PercentagePanel).